EMO-591 fix app server dependency direction
