+++ crates/verlet-kernel/src/cli/tool.rs
@@ -1666,8 +1666,8 @@ pub(crate) async fn open_mcp_source_registry(
     )
     .await
     .map_err(|err| {
-        if crate::cli::secret::turso_cross_process_lock_error(&err.to_string()) {
-            crate::cli::secret::cross_process_database_guidance(
+        if crate::adapters::app_server::instance::turso_cross_process_lock_error(&err.to_string()) {
+            crate::adapters::app_server::instance::cross_process_database_guidance(
                 "use the running daemon's mcpSource RPC or stop the daemon and retry",
             )
         } else {
